refactor(server): split server.go and webui.go into focused files
server.go (1650 lines) and webui.go (930 lines) had grown too large.
Split each by responsibility, keeping behaviour and exported surface
intact (pure file-move refactor — no logic changes).

server.go is now split across:
- server.go: Server struct, NewServer, watcher, adviser setup
- server_options.go: ServerOption + all With* + recording helpers
- server_runtime_flags.go: guardrails init + MCP feature flag helpers
- server_routes.go: middleware + route registration
- server_oauth_callback.go: dynamic OAuth callback server lifecycle
- server_lifecycle.go: Start/Stop/RemoteCoder/Codex undo

webui.go is now split across:
- webui.go: globals, UseUIEndpoints, static serving, audit sink
- webui_api.go: useWebAPIEndpoints + useV2Provider
- webui_auth.go: token validate/get/reset handlers
- webui_server_control.go: status/start/stop/restart/history handlers
