+++ internal/server/server.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"net"
 	"net/http"
 	"path/filepath"
 	"strings"
@@ -12,7 +11,6 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/pkg/browser"
 	"github.com/sirupsen/logrus"
 	"github.com/tingly-dev/tingly-box/ai"
 	"github.com/tingly-dev/tingly-box/ai/oauth"
@@ -21,8 +19,6 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/data"
 	"github.com/tingly-dev/tingly-box/internal/data/db"
 	"github.com/tingly-dev/tingly-box/internal/guardrails"
-	guardrailscore "github.com/tingly-dev/tingly-box/internal/guardrails/core"
-	guardrailsevaluate "github.com/tingly-dev/tingly-box/internal/guardrails/evaluate"
 	guardrailsutils "github.com/tingly-dev/tingly-box/internal/guardrails/utils"
 	"github.com/tingly-dev/tingly-box/internal/loadbalance"
 	mcpruntime "github.com/tingly-dev/tingly-box/internal/mcp/runtime"
@@ -33,7 +29,6 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/server/config"
 	"github.com/tingly-dev/tingly-box/internal/server/hooks"
 	"github.com/tingly-dev/tingly-box/internal/server/middleware"
-	"github.com/tingly-dev/tingly-box/internal/server/module/codeximport"
 	imbotmodule "github.com/tingly-dev/tingly-box/internal/server/module/imbot"
 	oauthmodule "github.com/tingly-dev/tingly-box/internal/server/module/oauth"
 	providerQuotaModule "github.com/tingly-dev/tingly-box/internal/server/module/provider_quota"
@@ -42,7 +37,6 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/server/servertool"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 	"github.com/tingly-dev/tingly-box/pkg/auth"
-	"github.com/tingly-dev/tingly-box/pkg/network"
 	pkgobs "github.com/tingly-dev/tingly-box/pkg/obs"
 	pkgotel "github.com/tingly-dev/tingly-box/pkg/otel"
 	"github.com/tingly-dev/tingly-box/pkg/otel/tracker"
@@ -197,308 +191,6 @@ func (s *Server) UsageStore() *db.UsageStore {
 // GetRoutingSelector returns the server's routing selector for service selection.
 func (s *Server) GetRoutingSelector() *routing.SimpleSelector {
 	return s.routingSelector
-}
-
-func (s *Server) initGuardrailsRuntime() {
-	runtime := s.currentGuardrailsRuntime()
-	if (runtime != nil && runtime.PolicyEngine() != nil) || s.config == nil {
-		return
-	}
-
-	if !s.guardrailsEnabled() {
-		return
-	}
-
-	cfgPath, err := FindGuardrailsConfig(s.config.ConfigDir)
-	if err != nil {
-		if !strings.Contains(err.Error(), "no guardrails config") {
-			logrus.WithError(err).Warn("Failed to locate guardrails config")
-			return
-		}
-		cfgPath, err = s.ensureDefaultGuardrailsConfig()
-		if err != nil {
-			logrus.WithError(err).Warn("Failed to create default guardrails config")
-			return
-		}
-	}
-
-	cfg, err := guardrails.LoadConfig(cfgPath)
-	if err != nil {
-		logrus.WithError(err).Warn("Failed to load guardrails config")
-		return
-	}
-
-	policy, err := guardrailsevaluate.BuildPolicyEngine(cfg, guardrailsevaluate.Dependencies{})
-	if err != nil {
-		logrus.WithError(err).Warn("Failed to build guardrails policy engine")
-		return
-	}
-
-	s.setGuardrailsRuntime(&guardrails.Guardrails{Policy: policy}, "guardrails init")
-	logrus.Infof("Guardrails enabled with config: %s", cfgPath)
-}
-
-func (s *Server) ensureDefaultGuardrailsConfig() (string, error) {
-	if s == nil || s.config == nil || s.config.ConfigDir == "" {
-		return "", fmt.Errorf("config directory not set")
-	}
-
-	path := GetGuardrailsConfigPath(s.config.ConfigDir)
-	cfg := guardrailscore.Config{
-		Groups: []guardrailscore.PolicyGroup{
-			{
-				ID:      guardrailscore.DefaultPolicyGroupID,
-				Name:    "Default",
-				Enabled: true,
-			},
-		},
-	}
-
-	data, err := marshalGuardrailsConfig(cfg)
-	if err != nil {
-		return "", err
-	}
-	if err := writeFileAtomic(path, data); err != nil {
-		return "", err
-	}
-
-	logrus.Infof("Created default guardrails config: %s", path)
-	return path, nil
-}
-
-func (s *Server) guardrailsEnabled() bool {
-	if s.config == nil {
-		return false
-	}
-	return s.config.GetScenarioFlag(typ.ScenarioGlobal, "guardrails") ||
-		s.config.GetScenarioFlag(typ.ScenarioClaudeCode, "guardrails")
-}
-
-// mcpEnabled checks if MCP feature is enabled via scenario flag
-func (s *Server) mcpEnabled() bool {
-	if s.config == nil {
-		return false
-	}
-	return s.config.GetScenarioFlag(typ.ScenarioGlobal, "mcp") ||
-		s.config.GetScenarioFlag(typ.ScenarioClaudeCode, "mcp")
-}
-
-// mcpStripDisabledToolsEnabled returns whether dangerous disabled MCP strip is enabled.
-func (s *Server) mcpStripDisabledToolsEnabled() bool {
-	if s.config == nil {
-		return false
-	}
-	cfg := s.config.GetMCPRuntimeConfig()
-	if cfg == nil {
-		return false
-	}
-	return cfg.StripDisabledMCPTools
-}
-
-// mcpMode returns the current MCP runtime mode
-func (s *Server) mcpMode() typ.MCPMode {
-	if s.config == nil {
-		return ""
-	}
-	var mcpCfg typ.MCPRuntimeConfig
-	if s.config.GetToolConfig(db.ToolTypeMCPRuntime, &mcpCfg) {
-		if mcpCfg.Mode == "" {
-			return typ.MCPModeClienttool // default mode
-		}
-		return mcpCfg.Mode
-	}
-	return typ.MCPModeClienttool // default mode
-}
-
-func (s *Server) syncGuardrailsFromConfig() {
-	if s.config == nil {
-		return
-	}
-
-	if !s.guardrailsEnabled() {
-		s.setGuardrailsRuntime(&guardrails.Guardrails{}, "guardrails disable")
-		logrus.Debug("Guardrails disabled via config")
-		return
-	}
-
-	if s.currentGuardrailsRuntime() == nil {
-		s.initGuardrailsRuntime()
-	}
-}
-
-// ServerOption defines a functional option for Server configuration
-type ServerOption func(*Server)
-
-// WithDefault applies all default server options
-func WithDefault() ServerOption {
-	return func(s *Server) {
-		s.enableUI = true      // Default: UI enabled
-		s.enableAdaptor = true // Default: adapter enabled
-		s.openBrowser = true   // Default: open browser enabled
-		s.host = ""            // Default: empty host (resolves to localhost)
-	}
-}
-
-func WithVersion(version string) ServerOption {
-	return func(s *Server) {
-		s.version = version
-	}
-}
-
-// WithUI enables or disables the UI for the server
-func WithUI(enabled bool) ServerOption {
-	return func(s *Server) {
-		s.enableUI = enabled
-	}
-}
-
-func WithHost(host string) ServerOption {
-	return func(s *Server) {
-		s.host = host
-	}
-}
-
-// WithAdaptor enables or disables the adaptor for the server
-func WithAdaptor(enabled bool) ServerOption {
-	return func(s *Server) {
-		s.enableAdaptor = enabled
-	}
-}
-
-// WithOpenBrowser enables or disables automatic browser opening
-func WithOpenBrowser(enabled bool) ServerOption {
-	return func(s *Server) {
-		s.openBrowser = enabled
-	}
-}
-
-// WithRecordMode sets the record mode for request/response recording
-// mode: empty string = disabled, "all" = record all, "response" = response only, "scenario" = record scenario only
-func WithRecordMode(mode obs.RecordMode) ServerOption {
-	return func(s *Server) {
-		s.recordMode = mode
-	}
-}
-
-// WithRecordDir sets the scenario-level record directory
-func WithRecordDir(dir string) ServerOption {
-	return func(s *Server) {
-		s.recordDir = dir
-	}
-}
-
-// WithRecordingCAS toggles content-addressed dedup alongside the default
-// gzip recording. When enabled, each session is written twice: once as a
-// gzip JSONL.gz (default), and once as content-addressed slim JSONL plus a
-// per-record blob tree. Useful for cross-session prompt analysis and replay.
-func WithRecordingCAS(enabled bool) ServerOption {
-	return func(s *Server) {
-		s.recordCAS = enabled
-	}
-}
-
-// WithRecording enables dual-stage recording for protocol conversion scenarios
-func WithRecording(enabled bool) ServerOption {
-	return func(s *Server) {
-		s.enableRecording = enabled
-	}
-}
-
-// WithGuardrails sets a guardrails runtime for stream evaluation.
-func WithGuardrails(runtime *guardrails.Guardrails) ServerOption {
-	return func(s *Server) {
-		s.setGuardrailsRuntimeRef(runtime)
-	}
-}
-
-// WithExperimentalFeatures sets the experimental features for the server
-func WithExperimentalFeatures(features map[string]bool) ServerOption {
-	return func(s *Server) {
-		s.experimentalFeatures = features
-	}
-}
-
-// WithDebug enables or disables debug mode for the server
-func WithDebug(enabled bool) ServerOption {
-	return func(s *Server) {
-		s.debug = enabled
-	}
-}
-
-// WithMultiLogger sets the multi-mode logger for the server
-func WithMultiLogger(logger *pkgobs.MultiLogger) ServerOption {
-	return func(s *Server) {
-		s.multiLogger = logger
-	}
-}
-
-// WithTemplateManager allows TBE to inject a custom TemplateManager.
-// This follows the same pattern as WithAuthMiddleware for consistency.
-func WithTemplateManager(tm *data.TemplateManager) ServerOption {
-	return func(s *Server) {
-		s.templateManager = tm
-		if s.config != nil {
-			s.config.SetTemplateManager(tm)
-		}
-	}
-}
-
-// IsFeatureEnabled checks if a specific feature is enabled
-func (s *Server) IsFeatureEnabled(feature string) bool {
-	return s.experimentalFeatures[feature]
-}
-
-// sinkOpts derives obs.SinkOption values from the server's recording config.
-func (s *Server) sinkOpts() []obs.SinkOption {
-	var opts []obs.SinkOption
-	if s.recordCAS {
-		opts = append(opts, obs.WithCASExporter())
-	}
-	return opts
-}
-
-// GetOrCreateScenarioSink gets or creates a recording sink for the specified scenario
-// The sink is created on-demand and cached for subsequent use
-func (s *Server) GetOrCreateScenarioSink(scenario typ.RuleScenario) *obs.Sink {
-	s.scenarioRecordSinksMu.Lock()
-	defer s.scenarioRecordSinksMu.Unlock()
-
-	// Return existing sink if already created
-	if sink, exists := s.scenarioRecordSinks[scenario]; exists {
-		return sink
-	}
-
-	mode := s.GetScenarioRecordMode(scenario)
-	if mode == "" {
-		return nil
-	}
-
-	// Create new sink for this scenario using the scenario-effective recording mode.
-	// This allows `recording_v2` on individual scenarios such as `claude_code`
-	// even when global CLI record mode is unset.
-	sink := obs.NewSink(s.recordDir, mode, s.sinkOpts()...)
-	if sink == nil {
-		// Sink creation failed or recording is disabled (empty recordDir)
-		// This is expected when no record directory is configured
-		logrus.Debugf("Failed to create scenario recording sink for %s", scenario)
-		return nil
-	}
-
-	s.scenarioRecordSinks[scenario] = sink
-	logrus.Debugf("Created scenario recording sink for %s, mode: %s, directory: %s", scenario, mode, s.recordDir)
-	return sink
-}
-
-func (s *Server) GetScenarioRecordMode(scenario typ.RuleScenario) obs.RecordMode {
-	if s == nil || s.config == nil {
-		return s.recordMode
-	}
-
-	if mode := s.config.GetScenarioRecordingMode(scenario); mode != typ.RecordingModeDisabled {
-		return obs.RecordMode(mode)
-	}
-
-	return s.recordMode
 }
 
 // NewServer creates a new HTTP server instance with functional options
@@ -912,739 +604,3 @@ func (s *Server) setupConfigWatcher() {
 	})
 }
 
-// startDynamicCallbackServer starts a temporary callback server for a specific OAuth session
-func (s *Server) startDynamicCallbackServer(sessionID string, port int) error {
-	s.callbackServersMu.Lock()
-	defer s.callbackServersMu.Unlock()
-
-	// Check if a callback server already exists for this session
-	if _, exists := s.callbackServers[sessionID]; exists {
-		return fmt.Errorf("callback server already exists for session %s", sessionID)
-	}
-
-	// Providers with fixed callback ports (for example Codex/OpenAI on 1455)
-	// can only have one active browser flow at a time. If a previous attempt
-	// is still waiting or failed before cleanup, replace it so retrying OAuth
-	// does not require waiting for the five minute timeout.
-	if port > 0 {
-		for existingSessionID, existingServer := range s.callbackServers {
-			if existingServer.GetPort() != port {
-				continue
-			}
-
-			_ = s.oauthManager.UpdateSessionStatus(existingSessionID, oauth.SessionStatusFailed, "", "Superseded by a new OAuth authorization attempt")
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			if err := existingServer.Stop(ctx); err != nil {
-				cancel()
-				return fmt.Errorf("failed to stop existing callback server on port %d: %w", port, err)
-			}
-			cancel()
-			delete(s.callbackServers, existingSessionID)
-			logrus.Debugf("[OAuth] Replaced existing dynamic callback server on port %d for session %s", port, existingSessionID)
-		}
-	}
-
-	// Create an http.HandlerFunc that properly handles the OAuth callback
-	handlerFunc := func(w http.ResponseWriter, r *http.Request) {
-		// Ignore favicon requests
-		if r.URL.Path == "/favicon.ico" {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		}
-
-		logrus.Debugf("[OAuth] Callback received: %s %s", r.Method, r.URL.Path)
-		logrus.Debugf("[OAuth] Query params: %v", r.URL.Query())
-
-		callbackOpts := oauthmodule.OAuthOptionsForSession(s.oauthManager, sessionID, fmt.Sprintf("http://localhost:%d", port))
-
-		// Delegate to the OAuth callback handler
-		// We need to directly call the oauth manager since gin won't work here
-		token, err := s.oauthManager.HandleCallback(r.Context(), r, callbackOpts...)
-		if err != nil {
-			logrus.Debugf("[OAuth] Callback error: %v", err)
-			if sessionID != "" {
-				_ = s.oauthManager.UpdateSessionStatus(sessionID, oauth.SessionStatusFailed, "", err.Error())
-				s.stopDynamicCallbackServerAfterResponse(sessionID)
-			}
-			w.Header().Set("Content-Type", "text/html")
-			w.WriteHeader(http.StatusBadRequest)
-			fmt.Fprintf(w, "<html><body><h1>OAuth Error</h1><p>%s</p></body></html>", err.Error())
-			return
-		}
-
-		// Use oauth handler to create the provider
-		providerUUID, err := s.oauthHandler.CreateProviderFromToken(token, token.Provider, "", token.SessionID)
-		if err != nil {
-			logrus.Debugf("[OAuth] Failed to create provider: %v", err)
-			failedSessionID := token.SessionID
-			if failedSessionID == "" {
-				failedSessionID = sessionID
-			}
-			if failedSessionID != "" {
-				_ = s.oauthManager.UpdateSessionStatus(failedSessionID, oauth.SessionStatusFailed, "", err.Error())
-			}
-			s.stopDynamicCallbackServerAfterResponse(sessionID)
-			w.Header().Set("Content-Type", "text/html")
-			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Fprintf(w, "<html><body><h1>OAuth Error</h1><p>Failed to create provider: %v</p></body></html>", err)
-			return
-		}
-
-		// Update session status to success if session ID exists
-		if token.SessionID != "" {
-			_ = s.oauthManager.UpdateSessionStatus(token.SessionID, oauth.SessionStatusSuccess, providerUUID, "")
-		}
-
-		logrus.Debugf("[OAuth] Callback successful for provider %s, created provider %s", token.Provider, providerUUID)
-
-		// Stop the dynamic callback server after successful callback
-		s.stopDynamicCallbackServerAfterResponse(sessionID)
-
-		// Return success HTML page
-		w.Header().Set("Content-Type", "text/html")
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprintf(w, `<!DOCTYPE html>
-<html>
-<head>
-    <title>OAuth Success</title>
-    <style>
-        body { font-family: Arial, sans-serif; max-width: 800px; margin: 50px auto; padding: 20px; }
-        .success { background: #d4edda; border: 1px solid #c3e6cb; color: #155724; padding: 20px; border-radius: 5px; }
-    </style>
-</head>
-<body>
-    <div class="success">
-        <h1>OAuth Authorization Successful!</h1>
-        <p>You can close this window and return to the application.</p>
-        <h2>Provider: %s</h2>
-        <p>Token: %s</p>
-    </div>
-</body>
-</html>`, string(token.Provider), oauthmodule.SafeTokenPreview(token.AccessToken))
-	}
-
-	// Create a new callback server with the handler
-	callbackServer := oauth.NewCallbackServer(handlerFunc)
-
-	// Start the callback server on the specified port
-	if err := callbackServer.Start(port); err != nil {
-		return fmt.Errorf("failed to start callback server on port %d: %w", port, err)
-	}
-
-	// Store the callback server reference
-	s.callbackServers[sessionID] = callbackServer
-
-	logrus.Debugf("[OAuth] Started dynamic callback server on port %d for session %s", port, sessionID)
-
-	// Auto-shutdown after 5 minutes
-	go func() {
-		time.Sleep(5 * time.Minute)
-		s.stopDynamicCallbackServer(sessionID)
-	}()
-
-	return nil
-}
-
-func (s *Server) stopDynamicCallbackServerAfterResponse(sessionID string) {
-	go func() {
-		time.Sleep(1 * time.Second) // Give time for the response to be sent
-		s.stopDynamicCallbackServer(sessionID)
-	}()
-}
-
-// stopDynamicCallbackServer stops and removes a dynamic callback server
-func (s *Server) stopDynamicCallbackServer(sessionID string) {
-	s.callbackServersMu.Lock()
-	defer s.callbackServersMu.Unlock()
-
-	callbackServer, exists := s.callbackServers[sessionID]
-	if !exists {
-		return
-	}
-
-	// Shutdown the callback server
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	if err := callbackServer.Stop(ctx); err != nil {
-		logrus.Debugf("[OAuth] Error stopping callback server for session %s: %v", sessionID, err)
-	}
-
-	// Remove from map
-	delete(s.callbackServers, sessionID)
-
-	logrus.Debugf("[OAuth] Stopped dynamic callback server for session %s", sessionID)
-}
-
-// StartDynamicCallbackServer starts a temporary callback server for OAuth
-// Implements CallbackServerManager interface for oauth module
-func (s *Server) StartDynamicCallbackServer(sessionID string, port int) error {
-	return s.startDynamicCallbackServer(sessionID, port)
-}
-
-// StopDynamicCallbackServer stops a temporary callback server for OAuth
-// Implements CallbackServerManager interface for oauth module
-func (s *Server) StopDynamicCallbackServer(sessionID string) {
-	s.stopDynamicCallbackServer(sessionID)
-}
-
-// setupMiddleware configures server middleware
-func (s *Server) setupMiddleware() {
-	// Recovery middleware
-	s.engine.Use(gin.Recovery())
-
-	// Memory log middleware for HTTP request logging
-	if s.memoryLogMW != nil {
-		s.engine.Use(s.memoryLogMW.Middleware())
-	}
-
-	// Debug middleware for logging requests/responses (only if enabled)
-	if s.errorMW != nil {
-		s.engine.Use(s.errorMW.Middleware())
-	}
-
-	// CORS middleware
-	s.engine.Use(middleware.CORS())
-}
-
-// setupRoutes configures server routes
-func (s *Server) setupRoutes(ctx context.Context) {
-
-	s.UseAIEndpoints()
-
-	s.UseLoadBalanceEndpoints()
-
-	// Multi-tenant token management API
-	s.UseTokenManagementEndpoints()
-
-	// Virtual model endpoints for testing
-	s.UseVirtualModelEndpoints()
-
-	// Integrate Web UI routes if enabled
-	if s.enableUI {
-		s.UseUIEndpoints(ctx)
-	}
-}
-
-func (s *Server) UseAIEndpoints() {
-	// DEPRECATED: now we only use path with scenario for openai and anthropic
-	//// OpenAI v1 API group
-	//openaiV1 := s.engine.Group("/openai/v1")
-	//s.SetupOpenAIEndpoints(openaiV1)
-	//
-	//// OpenAI API alias (without version)
-	//openai := s.engine.Group("/openai")
-	//s.SetupOpenAIEndpoints(openai)
-	//
-	//// Anthropic v1 API group
-	//anthropicV1 := s.engine.Group("/anthropic/v1")
-	//s.SetupAnthropicEndpoints(anthropicV1)
-
-	// Passthrough endpoints (no request/response transformation, just model replacement)
-	// Non-versioned passthrough routes
-	passthroughOpenai := s.engine.Group("/passthrough/openai")
-	s.SetupPassthroughOpenAIEndpoints(passthroughOpenai)
-
-	passthroughAnthropic := s.engine.Group("/passthrough/anthropic")
-	s.SetupPassthroughAnthropicEndpoints(passthroughAnthropic)
-
-	// Versioned passthrough routes
-	passthroughOpenaiV1 := s.engine.Group("/passthrough/openai/v1")
-	s.SetupPassthroughOpenAIEndpoints(passthroughOpenaiV1)
-
-	// scenario routes with middleware to inject scenario into context
-	scenario := s.engine.Group("/tingly/:scenario")
-	scenario.Use(s.contextMiddleware)
-	s.SetupMixinEndpoints(scenario)
-
-	// scenario v1 routes with middleware
-	scenarioV1 := s.engine.Group("/tingly/:scenario/v1")
-	scenarioV1.Use(s.contextMiddleware)
-	s.SetupMixinEndpoints(scenarioV1)
-}
-
-func (s *Server) SetupMixinEndpoints(group *gin.RouterGroup) {
-	// Chat completions endpoint (OpenAI compatible)
-	group.POST("/chat/completions", s.getModelAuthMiddleware(), s.HandleOpenAIChatCompletions)
-
-	// Responses API endpoints (OpenAI compatible)
-	group.POST("/responses", s.getModelAuthMiddleware(), s.HandleResponsesCreate)
-	group.GET("/responses/:id", s.getModelAuthMiddleware(), s.ResponsesGet)
-
-	// Chat completions endpoint (Anthropic compatible)
-	group.POST("/messages", s.getModelAuthMiddleware(), s.HandleAnthropicMessages)
-	// Count tokens endpoint (Anthropic compatible)
-	group.POST("/messages/count_tokens", s.getModelAuthMiddleware(), s.AnthropicCountTokens)
-
-	// Embeddings endpoint (OpenAI compatible)
-	group.POST("/embeddings", s.getModelAuthMiddleware(), s.HandleOpenAIEmbeddings)
-
-	// Image generation endpoint (OpenAI compatible).
-	// Routed directly to upstream POST /v1/images/generations; the Responses API
-	// (POST /responses with the image_generation tool) is exposed in parallel via
-	// the same scenario, with the caller choosing which surface to use.
-	group.POST("/images/generations", s.getModelAuthMiddleware(), s.HandleOpenAIImageGeneration)
-
-	// Models endpoint (routed by scenario: openai -> OpenAIListModels, anthropic/claude_code -> AnthropicListModels)
-	group.GET("/models", s.getModelAuthMiddleware(), s.ListModelsByScenario)
-}
-
-func (s *Server) SetupOpenAIEndpoints(group *gin.RouterGroup) {
-	// Chat completions endpoint (OpenAI compatible)
-	group.POST("/chat/completions", s.getModelAuthMiddleware(), s.HandleOpenAIChatCompletions)
-	// Models endpoint (OpenAI compatible)
-	group.GET("/models", s.getModelAuthMiddleware(), s.OpenAIListModels)
-
-	// Responses API endpoints (OpenAI compatible)
-	group.POST("/responses", s.getModelAuthMiddleware(), s.HandleResponsesCreate)
-	group.GET("/responses/:id", s.getModelAuthMiddleware(), s.ResponsesGet)
-}
-
-func (s *Server) SetupAnthropicEndpoints(group *gin.RouterGroup) {
-	// Chat completions endpoint (Anthropic compatible)
-	group.POST("/messages", s.getModelAuthMiddleware(), s.HandleAnthropicMessages)
-	// Count tokens endpoint (Anthropic compatible)
-	group.POST("/messages/count_tokens", s.getModelAuthMiddleware(), s.AnthropicCountTokens)
-	// Models endpoint (Anthropic compatible)
-	group.GET("/models", s.getModelAuthMiddleware(), s.AnthropicListModels)
-}
-
-// SetupPassthroughOpenAIEndpoints sets up pass-through endpoints for OpenAI-style requests
-// These endpoints bypass request/response transformations and only replace the model name
-func (s *Server) SetupPassthroughOpenAIEndpoints(group *gin.RouterGroup) {
-	// POST endpoints that use passthrough (proxy with model replacement)
-	group.POST("/chat/completions", s.getModelAuthMiddleware(), s.PassthroughOpenAI)
-	group.POST("/responses", s.getModelAuthMiddleware(), s.PassthroughOpenAI)
-	// GET responses/:id also uses passthrough
-	group.GET("/responses/*path", s.getModelAuthMiddleware(), s.PassthroughOpenAI)
-	// Models endpoint returns tingly-box's model list (not passthrough)
-	group.GET("/models", s.getModelAuthMiddleware(), s.OpenAIListModels)
-}
-
-// contextMiddleware is a middleware that extracts the scenario parameter from the URL path
-// and injects it into the request context for use by downstream components (e.g., RecordRoundTripper).
-// It also validates profile suffixes (e.g., "claude_code:p1") if present.
-func (s *Server) contextMiddleware(c *gin.Context) {
-	rawScenario := c.Param("scenario")
-	ctx := context.WithValue(c.Request.Context(), client.ScenarioContextKey, rawScenario)
-	c.Request = c.Request.WithContext(ctx)
-
-	// Validate profile if present (e.g., "claude_code:p1")
-	if typ.IsProfiledScenario(typ.RuleScenario(rawScenario)) {
-		base, profileID := typ.ParseScenarioProfile(typ.RuleScenario(rawScenario))
-		if base == "" || profileID == "" {
-			c.JSON(http.StatusBadRequest, gin.H{
-				"success": false,
-				"error":   fmt.Sprintf("invalid scenario format: '%s'", rawScenario),
-			})
-			c.Abort()
-			return
-		}
-
-		// Check base scenario exists in registry
-		if _, ok := typ.GetScenarioDescriptor(typ.RuleScenario(rawScenario)); !ok {
-			c.JSON(http.StatusBadRequest, gin.H{
-				"success": false,
-				"error":   fmt.Sprintf("unknown scenario '%s'", base),
-			})
-			c.Abort()
-			return
-		}
-
-		// Check profile exists in config
-		if s.config != nil {
-			if _, ok := s.config.GetProfile(typ.RuleScenario(base), profileID); !ok {
-				c.JSON(http.StatusBadRequest, gin.H{
-					"success": false,
-					"error":   fmt.Sprintf("unknown profile '%s' for scenario '%s'", profileID, base),
-				})
-				c.Abort()
-				return
-			}
-		}
-	}
-
-	c.Next()
-}
-
-// SetupPassthroughAnthropicEndpoints sets up pass-through endpoints for Anthropic-style requests
-// These endpoints bypass request/response transformations and only replace the model name
-func (s *Server) SetupPassthroughAnthropicEndpoints(group *gin.RouterGroup) {
-	// POST endpoints that use passthrough (proxy with model replacement)
-	group.POST("/messages", s.getModelAuthMiddleware(), s.PassthroughAnthropic)
-	group.POST("/messages/count_tokens", s.getModelAuthMiddleware(), s.PassthroughAnthropic)
-	// Models endpoint returns tingly-box's model list (not passthrough)
-	group.GET("/models", s.getModelAuthMiddleware(), s.AnthropicListModels)
-}
-
-// UseVirtualModelEndpoints sets up the direct virtual-model entrypoints,
-// split per protocol:
-//
-//	/virtual/openai/v1/{models,chat/completions}
-//	/virtual/anthropic/v1/{models,messages}
-//
-// These bypass the provider/rule/scenario pipeline and call the in-process
-// handler directly — useful when a client wants a fixed URL pointed at the
-// vmodel registry without configuring a provider. The protocol split
-// ensures /models returns only the model IDs the chosen protocol can
-// actually dispatch.
-//
-// The canonical path for virtual models in normal use is still
-// /v1/messages and /v1/chat/completions, where the dispatcher
-// short-circuits to the same handler when it resolves to a vmodel provider
-// (see HandleAnthropicMessages and HandleOpenAIChatCompletions).
-func (s *Server) UseVirtualModelEndpoints() {
-	mw := s.getModelAuthMiddleware()
-
-	openai := s.engine.Group("/virtual/openai")
-	openai.Use(mw)
-	s.virtualModelService.SetupOpenAIRoutes(openai)
-
-	anthropic := s.engine.Group("/virtual/anthropic")
-	anthropic.Use(mw)
-	s.virtualModelService.SetupAnthropicRoutes(anthropic)
-}
-
-func (s *Server) UseLoadBalanceEndpoints() {
-	// API routes for load balancer management
-	api := s.engine.Group("/api/v1/load-balancer")
-	api.Use(s.getUserAuthMiddleware()) // Require user authentication for management APIs
-
-	// Load balancer API routes
-	s.loadBalancerAPI.RegisterRoutes(api)
-}
-
-// UseTokenManagementEndpoints registers the token management API endpoints
-func (s *Server) UseTokenManagementEndpoints() {
-	// API routes for token management
-	api := s.engine.Group("/api/v1")
-	api.Use(s.getUserAuthMiddleware()) // Require user authentication for management APIs
-
-	// Token management API routes
-	s.registerTokenManagementAPI(api)
-}
-
-// Start starts the HTTP server
-func (s *Server) Start(port int) error {
-	// Start token refresher background goroutine
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	if s.oauthRefresher != nil {
-		go s.oauthRefresher.Start(ctx)
-		log.Println("OAuth token auto-refresh started")
-	}
-
-	// Start provider quota auto-refresh
-	if s.quotaManager != nil {
-		s.quotaManager.StartAutoRefresh(ctx)
-		log.Println("Provider quota auto-refresh started")
-	}
-
-	// Start configuration watcher
-	if s.watcher != nil {
-		if err := s.watcher.Start(); err != nil {
-			logrus.Debugf("Failed to start config watcher: %v", err)
-		} else {
-			log.Println("Configuration hot-reload enabled")
-		}
-	}
-
-	// Start remote coder service (auto-start by default)
-	if err := s.StartRemoteCoder(); err != nil {
-		logrus.WithError(err).Warn("Failed to auto-start remote-coder")
-	} else {
-		logrus.Info("Remote-coder auto-start initiated")
-	}
-
-	addr := fmt.Sprintf("%s:%d", s.host, port)
-	s.httpServer = &http.Server{
-		Addr:              addr,
-		Handler:           s.engine,
-		ReadHeaderTimeout: 10 * time.Second,
-		ReadTimeout:       30 * time.Second,
-		WriteTimeout:      10 * time.Minute,
-		IdleTimeout:       120 * time.Second,
-	}
-
-	resolvedHost := network.ResolveHost(s.host)
-	scheme := "http"
-
-	// CASE 1: Non-UI Mode ---
-	if !s.enableUI {
-		fmt.Printf("OpenAI v1 Chat API endpoint: %s://%s:%d/openai/v1/chat/completions\n", scheme, resolvedHost, port)
-		fmt.Printf("Anthropic v1 Message API endpoint: %s://%s:%d/anthropic/v1/messages\n", scheme, resolvedHost, port)
-		fmt.Printf("Embeddings API endpoint: %s://%s:%d/tingly/embed/v1/embeddings\n", scheme, resolvedHost, port)
-		fmt.Printf("Image Generation API endpoint: %s://%s:%d/tingly/imagegen/v1/images/generations\n", scheme, resolvedHost, port)
-		fmt.Printf("Image Generation (Responses API): %s://%s:%d/tingly/imagegen/v1/responses\n", scheme, resolvedHost, port)
-		fmt.Printf("Virtual Model API (OpenAI): %s://%s:%d/virtual/openai/v1/chat/completions\n", scheme, resolvedHost, port)
-		fmt.Printf("Virtual Model API (Anthropic): %s://%s:%d/virtual/anthropic/v1/messages\n", scheme, resolvedHost, port)
-		fmt.Printf("Mode name: %s\n", constant.DefaultModeName)
-		fmt.Printf("Model API key: %s\n", s.config.GetModelToken())
-		return s.httpServer.ListenAndServe()
-	}
-
-	// CASE 2: Web UI Mode ---
-	webUIURL := fmt.Sprintf("%s://%s:%d", scheme, resolvedHost, port)
-	if s.config.HasUserToken() {
-		webUIURL = fmt.Sprintf("%s/login/%s", webUIURL, s.config.GetUserToken())
-	}
-
-	fmt.Printf("Web UI: %s\n", webUIURL)
-	if s.openBrowser {
-		fmt.Printf("Starting server and opening browser...\n")
-	} else {
-		fmt.Printf("Starting server...\n")
-	}
-
-	// Use a channel to capture the immediate error if ListenAndServe fails
-	serverError := make(chan error, 1)
-	go func() {
-		serverError <- s.httpServer.ListenAndServe()
-	}()
-
-	// Instead of a fixed 100ms sleep, we poll the port
-	if err := waitForPort(addr, 2*time.Second); err != nil {
-		// Check if the server goroutine already caught a "port in use" error
-		select {
-		case e := <-serverError:
-			return e
-		default:
-			return fmt.Errorf("timeout: server did not start on %s: %v", addr, err)
-		}
-	}
-
-	// Server is up, now open browser if enabled
-	if s.openBrowser {
-		browser.OpenURL(webUIURL)
-	}
-
-	// Block until server shuts down or errors out
-	return <-serverError
-}
-
-// Helper: Polls the port to ensure it's open before browser opens
-func waitForPort(addr string, timeout time.Duration) error {
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		conn, err := net.DialTimeout("tcp", addr, 100*time.Millisecond)
-		if err == nil {
-			conn.Close()
-			return nil
-		}
-		time.Sleep(50 * time.Millisecond)
-	}
-	return fmt.Errorf("port %s not reachable", addr)
-}
-
-// GetRouter returns the Gin engine for testing purposes
-func (s *Server) GetRouter() *gin.Engine {
-	return s.engine
-}
-
-// GetLoadBalancer returns the load balancer instance
-func (s *Server) GetLoadBalancer() *LoadBalancer {
-	return s.loadBalancer
-}
-
-// HealthMonitor returns the server's health monitor
-func (s *Server) HealthMonitor() *loadbalance.HealthMonitor {
-	return s.healthMonitor
-}
-
-// StartRemoteCoder starts the remote control service if not already running
-func (s *Server) StartRemoteCoder() error {
-	s.remoteCoderMu.Lock()
-	defer s.remoteCoderMu.Unlock()
-
-	// Already running
-	if s.remoteCoderCancel != nil {
-		logrus.Debug("Remote control already running")
-		return nil
-	}
-
-	// Check if imbotsettings handler is available
-	if s.imbotSettingsHandler == nil {
-		return fmt.Errorf("imbotsettings handler not available")
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	s.remoteCoderCtx = ctx
-	s.remoteCoderCancel = cancel
-
-	logrus.Info("Starting remote control service...")
-
-	// Start all enabled bots through the imbotsettings handler
-	go func() {
-		if err := s.imbotSettingsHandler.StartAllEnabled(ctx); err != nil && ctx.Err() == nil {
-			logrus.WithError(err).Warn("Failed to start some enabled bots")
-		}
-		// Keep context alive until canceled
-		<-ctx.Done()
-		logrus.Info("Remote control service stopped")
-	}()
-
-	return nil
-}
-
-// StopRemoteCoder stops the remote control service if running
-func (s *Server) StopRemoteCoder() {
-	s.remoteCoderMu.Lock()
-	defer s.remoteCoderMu.Unlock()
-
-	if s.remoteCoderCancel == nil {
-		logrus.Debug("Remote control not running")
-		return
-	}
-
-	// Cancel the context first
-	s.remoteCoderCancel()
-	s.remoteCoderCancel = nil
-	s.remoteCoderCtx = nil
-
-	// Stop all bots through the imbotsettings handler
-	if s.imbotSettingsHandler != nil {
-		s.imbotSettingsHandler.StopAll()
-		logrus.Info("All bots stopped via imbotsettings handler")
-	}
-
-	logrus.Info("Remote-coder stopped")
-}
-
-// IsRemoteCoderRunning returns whether the remote control service is running
-func (s *Server) IsRemoteCoderRunning() bool {
-	s.remoteCoderMu.Lock()
-	defer s.remoteCoderMu.Unlock()
-	return s.remoteCoderCancel != nil
-}
-
-// SyncRemoteCoderBots syncs bots with the remote control bot manager
-func (s *Server) SyncRemoteCoderBots(ctx context.Context) error {
-	if s.imbotSettingsHandler == nil {
-		return fmt.Errorf("bot manager not available")
-	}
-	return s.imbotSettingsHandler.Sync(ctx)
-}
-
-// Stop gracefully stops the HTTP server
-func (s *Server) Stop(ctx context.Context) error {
-	if s.cancel != nil {
-		s.cancel()
-		s.cancel = nil
-		s.ctx = nil
-	}
-
-	if s.httpServer == nil {
-		return nil
-	}
-
-	// Stop remote control if running
-	s.StopRemoteCoder()
-
-	// Shutdown ImBot settings handler
-	if s.imbotSettingsHandler != nil {
-		s.imbotSettingsHandler.Shutdown()
-		log.Println("ImBot settings handler stopped")
-	}
-
-	// Stop token refresher
-	if s.oauthRefresher != nil {
-		s.oauthRefresher.Stop()
-		log.Println("OAuth token auto-refresh stopped")
-	}
-
-	// Stop provider quota auto-refresh
-	if s.quotaManager != nil {
-		s.quotaManager.StopAutoRefresh()
-		log.Println("Provider quota auto-refresh stopped")
-	}
-
-	if err := s.undoCodexImportOnStop(); err != nil {
-		logrus.WithError(err).Warn("Failed to auto-undo Codex import on stop")
-	}
-
-	// Stop debug middleware
-	if s.errorMW != nil {
-		s.errorMW.Stop()
-	}
-
-	// Stop configuration watcher
-	if s.watcher != nil {
-		s.watcher.Stop()
-		log.Println("Configuration watcher stopped")
-	}
-
-	// Close all MCP sessions and terminate subprocesses
-	if s.mcpRuntime != nil {
-		s.mcpRuntime.Close()
-		log.Println("MCP runtime stopped: all sessions closed and subprocesses terminated")
-	}
-
-	// Close all scenario recording sinks
-	s.scenarioRecordSinksMu.Lock()
-	for scenario, sink := range s.scenarioRecordSinks {
-		if sink != nil {
-			sink.Close()
-			logrus.Debugf("Closed scenario recording sink for %s", scenario)
-		}
-	}
-	s.scenarioRecordSinks = make(map[typ.RuleScenario]*obs.Sink)
-	s.scenarioRecordSinksMu.Unlock()
-
-	// Shutdown OTel meter setup
-	if s.meterSetup != nil {
-		if err := s.meterSetup.Shutdown(ctx); err != nil {
-			logrus.Errorf("OTel shutdown error: %v", err)
-		}
-	}
-
-	// Close all database stores via StoreManager
-	if s.config.StoreManager() != nil {
-		if err := s.config.StoreManager().Close(); err != nil {
-			logrus.Errorf("Error closing stores: %v", err)
-		}
-	}
-
-	fmt.Println("Shutting down server...")
-	return s.httpServer.Shutdown(ctx)
-}
-
-func (s *Server) undoCodexImportOnStop() error {
-	if s == nil || s.config == nil {
-		return nil
-	}
-	if !s.config.GetScenarioExtensionBool(typ.ScenarioCodex, codeximport.ImportStateAutoUndoOnStopKey()) {
-		return nil
-	}
-	if !s.config.GetScenarioExtensionBool(typ.ScenarioCodex, codeximport.ImportStateActiveKey()) {
-		return nil
-	}
-
-	targetProvider := s.config.GetScenarioExtensionString(typ.ScenarioCodex, codeximport.ImportStateSourceProviderKey())
-	sourceProvider := s.config.GetScenarioExtensionString(typ.ScenarioCodex, codeximport.ImportStateTargetProviderKey())
-	if sourceProvider == "" || targetProvider == "" {
-		return nil
-	}
-
-	importer := codeximport.NewImporter()
-	_, err := importer.ImportOpenAISessions(codeximport.ImportOpenAISessionsRequest{
-		SourceProvider: sourceProvider,
-		TargetProvider: targetProvider,
-		CodexHome:      s.config.GetScenarioExtensionString(typ.ScenarioCodex, codeximport.ImportStateCodexHomeKey()),
-		SqliteHome:     s.config.GetScenarioExtensionString(typ.ScenarioCodex, codeximport.ImportStateSqliteHomeKey()),
-		StateDBPath:    s.config.GetScenarioExtensionString(typ.ScenarioCodex, codeximport.ImportStateStateDBPathKey()),
-		DryRun:         false,
-	})
-	if err != nil {
-		return err
-	}
-	return s.config.SetScenarioExtensions(typ.ScenarioCodex, map[string]interface{}{
-		codeximport.ImportStateActiveKey():         false,
-		codeximport.ImportStateSourceProviderKey(): nil,
-		codeximport.ImportStateTargetProviderKey(): nil,
-		codeximport.ImportStateCodexHomeKey():      nil,
-		codeximport.ImportStateSqliteHomeKey():     nil,
-		codeximport.ImportStateStateDBPathKey():    nil,
-		codeximport.ImportStateAutoUndoOnStopKey(): false,
-	})
-}
